requirements.txt
megatron-core>=0.12
modelscope
peft>=0.11,<0.19
safetensors
tqdm
transformers>=4.33,<5.4.0
